# Minor optimizations in write_coeffs_lv_map