Up to 3% improvement observed on write_b benchmarks

```
write_b_bench(TX_4X4, 20)                                                                             
                        time:   [1.1021 ms 1.1056 ms 1.1093 ms]
                        change: [-3.4999% -2.8032% -2.1049%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

write_b_bench(TX_4X4, 55)                                                                             
                        time:   [1.0964 ms 1.0990 ms 1.1021 ms]
                        change: [-3.4298% -2.8635% -2.2903%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) high mild
  7 (7.00%) high severe

write_b_bench(TX_8X8, 20)                                                                            
                        time:   [560.50 us 562.63 us 565.22 us]
                        change: [-2.5384% -1.9839% -1.4317%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

write_b_bench(TX_8X8, 55)                                                                            
                        time:   [558.67 us 560.42 us 562.55 us]
                        change: [-2.5445% -1.8203% -1.2177%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
```